+++ projects/core/src/main/kotlin/site/siredvin/peripheralium/api/peripheral/IPluggablePeripheral.kt
@@ -1,9 +1,10 @@
 package site.siredvin.peripheralium.api.peripheral
 
 import dan200.computercraft.api.peripheral.IComputerAccess
+import dan200.computercraft.api.peripheral.IPeripheral
 import java.util.function.Consumer
 
-interface IPluggablePeripheral {
+interface IPluggablePeripheral: IPeripheral {
     fun forEachComputer(func: Consumer<IComputerAccess>)
     fun queueEvent(event: String, vararg arguments: Any) {
         forEachComputer {