+++ projects/core/src/main/kotlin/site/siredvin/peripheralium/computercraft/operations/SphereOperationContext.kt
@@ -1,0 +1,12 @@
+package site.siredvin.peripheralium.computercraft.operations
+
+import java.io.Serializable
+
+class SphereOperationContext(val radius: Int) : Serializable {
+
+    companion object {
+        fun of(radius: Int): SphereOperationContext {
+            return SphereOperationContext(radius)
+        }
+    }
+}