+++ projects/fabric/src/main/kotlin/site/siredvin/peripheralium/FabricPeripheralium.kt
@@ -2,6 +2,8 @@ package site.siredvin.peripheralium
 
 import fuzs.forgeconfigapiport.api.config.v2.ForgeConfigRegistry
 import net.fabricmc.api.ModInitializer
+import net.fabricmc.fabric.api.itemgroup.v1.FabricItemGroup
+import net.minecraft.resources.ResourceLocation
 import net.minecraftforge.fml.config.ModConfig
 import site.siredvin.peripheralium.api.storage.ExtractorProxy
 import site.siredvin.peripheralium.common.configuration.ConfigHolder
@@ -15,6 +17,9 @@ object FabricPeripheralium: ModInitializer {
 
     init {
         PeripheraliumCore.configure(FabricPeripheraliumPlatform(), FabricIngredients)
+        PeripheraliumCore.configureCreativeTab(FabricItemGroup.builder(
+            ResourceLocation(PeripheraliumCore.MOD_ID, "tab")
+        )).build()
         // Register extract storages
         ExtractorProxy.addStorageExtractor(FabricStorageUtils::extractStorage)
     }